change money types to BigDecimal

+++ src/main/kotlin/it/walletwap/ewallet/dto/TransactionDto.kt
@@ -1,7 +1,9 @@
 package it.walletwap.ewallet.dto
 
+import java.math.BigDecimal
+
 data class TransactionDto(
-    var amountTransferred: Long,
+    var amountTransferred: BigDecimal,
     var walletFromId: Long, var walletToId: Long
 )
