+++ src/main/kotlin/it/walletwap/ewallet/dto/WalletDto.kt
@@ -1,6 +1,8 @@
 package it.walletwap.ewallet.dto
 
+import java.math.BigDecimal
+
 data class WalletDto(
     var id: Long?,
-    var amount: Long
+    var amount: BigDecimal
 )
